# routes/admin/v1/admin.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "FG Digital",
        "description": "ADMIN APIs - Project Management Portal",
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/admin/v1",
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
...
    "paths": {
...
        "/client/get": {
            "get": {
                "tags": ["4. Client"],
                "summary": "Get All Client",
                "parameters": [{
                    "in": "query",
                    "name": "client_id",
                    "description": "Client ID",
                    "required": false,
                    "items": { "type": "string" }
                }],
                "responses": {}
            }

        },
...
        "/project/get": {
            "get": {
                "tags": ["6. Client-project"],
                "summary": "Get All Project Details",
                "parameters": [{
                    "in": "query",
                    "name": "client_id",
                    "description": "get project",
                    "required": false,
                    "items": { "type": "string" }
                }],
                "responses": {}
            }

        },
        "/project/team/add": {
            "post": {
                "tags": ["6. Client-project"],
                "summary": "Assign Team member to project",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["project_id", "client_id"],
                        "properties": {
                            "employee_id": { "type": "string" },
                            "project_id": { "type": "string" }
                        },
                        "example": {
                            "employee_id": "20201",
                            "project_id": "1376735095048"
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/project/team/update": {
            "post": {
                "tags": ["6. Client-project"],
                "summary": "Update Permission to team member",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["project_id", "client_id"],
                        "properties": {
                            "employee_id": { "type": "string" },
                            "project_id": { "type": "string" },
                            "active": { "type": "boolean" },
                            "review": { "type": "boolean" },
                            "activity": { "type": "boolean" }

                        },
                        "example": {
                            "employee_id": "20201",
                            "project_id": "1376735095048",
                            "active": "true"
                        }
                    }
                }],
                "responses": {}
            }
        },
...
    }
}